update/ merge wordIndexes
