--- constants/wordsIndex.tsx
@@ -1,11 +1,0 @@
-import { GroupTagItemType } from "./globalTypes";
-
-export const groupTags: GroupTagItemType[] = [
-  { id: "beginnerTag", text: "新手團" },
-  { id: "expertTag", text: "老手團" },
-  { id: "practiceTag", text: "經驗切磋" },
-  { id: "openTag", text: "不限定" },
-  { id: "tutorialTag", text: "教學團" },
-  { id: "casualTag", text: "輕鬆" },
-  { id: "competitiveTag", text: "競技" },
-];
